+++ changelog.md
@@ -1,6 +1,7 @@
 ## 2.0.0
 - Injecting components/mixins/directives is now optional
 - There is now a default merging strategy for when using mixins with dependencies
+- Added a `injector.strict` option which makes all dependencies optional. Note this doesn't affect `injector.get`
 
 ## 1.0.0
 - Upgraded to Jpex 2.0.0